front/syn_metadata: move type-string rendering helpers out of front/ast (Slice 2)
Lift full_type_string, qualified_full_type_string,
qualified_full_type_string_with_imports, qualify_known_trait_name,
trait_object_root_name, and trait_object_root_name_qualified from
front::ast into front::syn_metadata. parse.rs callsites and the
duplicated qualify_known_trait_name in parse.rs route through the new
home. front::ast keeps pub-use re-exports for the in-module callers.
extract_dyn_trait_root_with_context (still in front::ast) delegates
trait-object lookup to syn_metadata::trait_object_root_name_qualified.
